Added a class "Item" class, and three types of items
